+++ integration_tests/juniper_tests/src/codegen/impl_scalar.rs
@@ -9,6 +9,7 @@ struct DefaultName(i32);
 struct OtherOrder(i32);
 struct Named(i32);
 struct ScalarDescription(i32);
+struct ScalarSpecifiedByUrl(i32);
 struct Generated(String);
 
 struct Root;
@@ -31,10 +32,10 @@ where
         Value::scalar(self.0)
     }
 
-    fn from_input_value(v: &InputValue) -> Option<DefaultName> {
-        v.as_scalar_value()
-            .and_then(|s| s.as_int())
-            .map(|i| DefaultName(i))
+    fn from_input_value(v: &InputValue) -> Result<DefaultName, String> {
+        v.as_int_value()
+            .map(DefaultName)
+            .ok_or_else(|| format!("Expected `Int`, found: {}", v))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, S> {
@@ -48,8 +49,10 @@ impl GraphQLScalar for OtherOrder {
         Value::scalar(self.0)
     }
 
-    fn from_input_value(v: &InputValue) -> Option<OtherOrder> {
-        v.as_scalar_value::<i32>().map(|i| OtherOrder(*i))
+    fn from_input_value(v: &InputValue) -> Result<OtherOrder, String> {
+        v.as_int_value()
+            .map(OtherOrder)
+            .ok_or_else(|| format!("Expected `Int`, found: {}", v))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, DefaultScalarValue> {
@@ -63,8 +66,10 @@ impl GraphQLScalar for Named {
         Value::scalar(self.0)
     }
 
-    fn from_input_value(v: &InputValue) -> Option<Named> {
-        v.as_scalar_value::<i32>().map(|i| Named(*i))
+    fn from_input_value(v: &InputValue) -> Result<Named, String> {
+        v.as_int_value()
+            .map(Named)
+            .ok_or_else(|| format!("Expected `Int`, found: {}", v))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, DefaultScalarValue> {
@@ -78,8 +83,27 @@ impl GraphQLScalar for ScalarDescription {
         Value::scalar(self.0)
     }
 
-    fn from_input_value(v: &InputValue) -> Option<ScalarDescription> {
-        v.as_scalar_value::<i32>().map(|i| ScalarDescription(*i))
+    fn from_input_value(v: &InputValue) -> Result<ScalarDescription, String> {
+        v.as_int_value()
+            .map(ScalarDescription)
+            .ok_or_else(|| format!("Expected `Int`, found: {}", v))
+    }
+
+    fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, DefaultScalarValue> {
+        <i32 as ParseScalarValue>::from_str(value)
+    }
+}
+
+#[graphql_scalar(specified_by_url = "https://tools.ietf.org/html/rfc4122")]
+impl GraphQLScalar for ScalarSpecifiedByUrl {
+    fn resolve(&self) -> Value {
+        Value::scalar(self.0)
+    }
+
+    fn from_input_value(v: &InputValue) -> Result<ScalarSpecifiedByUrl, String> {
+        v.as_int_value()
+            .map(ScalarSpecifiedByUrl)
+            .ok_or_else(|| format!("Expected `Int`, found: {}", v))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, DefaultScalarValue> {
@@ -98,10 +122,11 @@ macro_rules! impl_scalar {
                 Value::scalar(self.0.clone())
             }
 
-            fn from_input_value(v: &InputValue) -> Option<Self> {
+            fn from_input_value(v: &InputValue) -> Result<Self, &'static str> {
                 v.as_scalar_value()
                     .and_then(|v| v.as_str())
                     .and_then(|s| Some(Self(s.to_owned())))
+                    .ok_or_else(|| "Expected `String`")
             }
 
             fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, S> {
@@ -127,6 +152,9 @@ impl Root {
     fn scalar_description() -> ScalarDescription {
         ScalarDescription(0)
     }
+    fn scalar_specified_by_url() -> ScalarSpecifiedByUrl {
+        ScalarSpecifiedByUrl(0)
+    }
     fn generated() -> Generated {
         Generated("foo".to_owned())
     }
@@ -140,9 +168,10 @@ impl GraphQLScalar for WithCustomScalarValue {
         Value::scalar(self.0)
     }
 
-    fn from_input_value(v: &InputValue<MyScalarValue>) -> Option<WithCustomScalarValue> {
-        v.as_scalar_value::<i32>()
-            .map(|i| WithCustomScalarValue(*i))
+    fn from_input_value(v: &InputValue<MyScalarValue>) -> Result<WithCustomScalarValue, String> {
+        v.as_int_value()
+            .map(WithCustomScalarValue)
+            .ok_or_else(|| format!("Expected Int, found: {}", v))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, MyScalarValue> {
@@ -198,8 +227,10 @@ fn path_in_resolve_return_type() {
             Value::scalar(self.0)
         }
 
-        fn from_input_value(v: &InputValue) -> Option<ResolvePath> {
-            v.as_scalar_value::<i32>().map(|i| ResolvePath(*i))
+        fn from_input_value(v: &InputValue) -> Result<ResolvePath, String> {
+            v.as_int_value()
+                .map(ResolvePath)
+                .ok_or_else(|| format!("Expected `Int`, found: {}", v))
         }
 
         fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, DefaultScalarValue> {
@@ -287,6 +318,7 @@ async fn scalar_description_introspection() {
         __type(name: "ScalarDescription") {
             name
             description
+            specifiedByUrl
         }
     }
     "#;
@@ -301,6 +333,32 @@ async fn scalar_description_introspection() {
             Some(&graphql_value!(
                 "A sample scalar, represented as an integer",
             )),
+        );
+        assert_eq!(
+            type_info.get_field_value("specifiedByUrl"),
+            Some(&graphql_value!(null)),
+        );
+    })
+    .await;
+}
+
+#[tokio::test]
+async fn scalar_specified_by_url_introspection() {
+    let doc = r#"{
+        __type(name: "ScalarSpecifiedByUrl") {
+            name
+            specifiedByUrl
+        }
+    }"#;
+
+    run_type_info_query(doc, |type_info| {
+        assert_eq!(
+            type_info.get_field_value("name"),
+            Some(&graphql_value!("ScalarSpecifiedByUrl")),
+        );
+        assert_eq!(
+            type_info.get_field_value("specifiedByUrl"),
+            Some(&graphql_value!("https://tools.ietf.org/html/rfc4122")),
         );
     })
     .await;