+++ juniper/src/executor_tests/introspection/input_object.rs
@@ -204,9 +204,9 @@ fn default_name_input_value() {
         "fieldTwo": "number two",
     });
 
-    let dv: Option<DefaultName> = FromInputValue::from_input_value(&iv);
+    let dv = DefaultName::from_input_value(&iv);
 
-    assert!(dv.is_some());
+    assert!(dv.is_ok(), "error: {}", dv.unwrap_err().message());
 
     let dv = dv.unwrap();