+++ juniper/src/integrations/mod.rs
@@ -8,6 +8,8 @@ pub mod chrono;
 pub mod chrono_tz;
 #[doc(hidden)]
 pub mod serde;
+#[cfg(feature = "time")]
+pub mod time;
 #[cfg(feature = "url")]
 pub mod url;
 #[cfg(feature = "uuid")]