+++ juniper/src/macros/mod.rs
@@ -1,5 +1,6 @@
 //! Declarative macros and helper definitions for procedural macros.
 
+#[doc(hidden)]
 pub mod helper;
 
 #[macro_use]