+++ juniper/src/tests/introspection_tests.rs
@@ -208,6 +208,19 @@ async fn test_introspection_directives() {
                         "INLINE_FRAGMENT",
                     ],
                 },
+                {
+                    "name": "deprecated",
+                    "locations": [
+                        "FIELD_DEFINITION",
+                        "ENUM_VALUE",
+                    ],
+                },
+                {
+                    "name": "specifiedBy",
+                    "locations": [
+                        "SCALAR",
+                    ],
+                },
             ],
         },
     });