+++ juniper/src/types/scalars.rs
@@ -59,14 +59,12 @@ where
         Value::scalar(self.0.clone())
     }
 
-    fn from_input_value(v: &InputValue) -> Option<ID> {
-        match *v {
-            InputValue::Scalar(ref s) => s
-                .as_string()
-                .or_else(|| s.as_int().map(|i| i.to_string()))
-                .map(ID),
-            _ => None,
-        }
+    fn from_input_value(v: &InputValue) -> Result<ID, String> {
+        v.as_string_value()
+            .map(str::to_owned)
+            .or_else(|| v.as_int_value().map(|i| i.to_string()))
+            .map(ID)
+            .ok_or_else(|| format!("Expected `String` or `Int`, found: {}", v))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, S> {
@@ -86,11 +84,10 @@ where
         Value::scalar(self.clone())
     }
 
-    fn from_input_value(v: &InputValue) -> Option<String> {
-        match *v {
-            InputValue::Scalar(ref s) => s.as_string(),
-            _ => None,
-        }
+    fn from_input_value(v: &InputValue) -> Result<String, String> {
+        v.as_string_value()
+            .map(str::to_owned)
+            .ok_or_else(|| format!("Expected `String`, found: {}", v))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, S> {
@@ -275,11 +272,10 @@ where
         Value::scalar(*self)
     }
 
-    fn from_input_value(v: &InputValue) -> Option<bool> {
-        match *v {
-            InputValue::Scalar(ref b) => b.as_boolean(),
-            _ => None,
-        }
+    fn from_input_value(v: &InputValue) -> Result<bool, String> {
+        v.as_scalar_value()
+            .and_then(ScalarValue::as_boolean)
+            .ok_or_else(|| format!("Expected `Boolean`, found: {}", v))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, S> {
@@ -297,11 +293,9 @@ where
         Value::scalar(*self)
     }
 
-    fn from_input_value(v: &InputValue) -> Option<i32> {
-        match *v {
-            InputValue::Scalar(ref i) => i.as_int(),
-            _ => None,
-        }
+    fn from_input_value(v: &InputValue) -> Result<i32, String> {
+        v.as_int_value()
+            .ok_or_else(|| format!("Expected `Int`, found: {}", v))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, S> {
@@ -324,11 +318,9 @@ where
         Value::scalar(*self)
     }
 
-    fn from_input_value(v: &InputValue) -> Option<f64> {
-        match *v {
-            InputValue::Scalar(ref s) => s.as_float(),
-            _ => None,
-        }
+    fn from_input_value(v: &InputValue) -> Result<f64, String> {
+        v.as_float_value()
+            .ok_or_else(|| format!("Expected `Float`, found: {}", v))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, S> {