+++ juniper_codegen/src/util/mod.rs
@@ -17,6 +17,7 @@ use syn::{
     spanned::Spanned,
     token, Attribute, Ident, Lit, Meta, MetaList, MetaNameValue, NestedMeta,
 };
+use url::Url;
 
 use crate::common::parse::ParseBufferExt as _;
 
@@ -454,6 +455,7 @@ pub enum FieldAttributeParseMode {
 enum FieldAttribute {
     Name(SpanContainer<syn::LitStr>),
     Description(SpanContainer<syn::LitStr>),
+    SpecifiedByUrl(SpanContainer<syn::LitStr>),
     Deprecation(SpanContainer<DeprecationAttr>),
     Skip(SpanContainer<syn::Ident>),
     Arguments(HashMap<String, FieldAttributeArgument>),
@@ -483,6 +485,15 @@ impl Parse for FieldAttribute {
                 input.parse::<token::Eq>()?;
                 let lit = input.parse::<syn::LitStr>()?;
                 Ok(FieldAttribute::Description(SpanContainer::new(
+                    ident.span(),
+                    Some(lit.span()),
+                    lit,
+                )))
+            }
+            "specified_by_url" => {
+                input.parse::<token::Eq>()?;
+                let lit = input.parse::<syn::LitStr>()?;
+                Ok(FieldAttribute::SpecifiedByUrl(SpanContainer::new(
                     ident.span(),
                     Some(lit.span()),
                     lit,
@@ -542,7 +553,9 @@ pub struct FieldAttributes {
     pub name: Option<SpanContainer<String>>,
     pub description: Option<SpanContainer<String>>,
     pub deprecation: Option<SpanContainer<DeprecationAttr>>,
-    // Only relevant for GraphQLObject derive.
+    /// Only relevant for scalar impl macro.
+    pub specified_by_url: Option<SpanContainer<Url>>,
+    /// Only relevant for GraphQLObject derive.
     pub skip: Option<SpanContainer<syn::Ident>>,
     /// Only relevant for object macro.
     pub arguments: HashMap<String, FieldAttributeArgument>,
@@ -563,6 +576,18 @@ impl Parse for FieldAttributes {
                 }
                 FieldAttribute::Description(name) => {
                     output.description = Some(name.map(|val| val.value()));
+                }
+                FieldAttribute::SpecifiedByUrl(url) => {
+                    output.specified_by_url = Some(
+                        url.map(|val| Url::parse(&val.value()))
+                            .transpose()
+                            .map_err(|e| {
+                                syn::Error::new(
+                                    e.span_ident(),
+                                    format!("Invalid URL: {}", e.inner()),
+                                )
+                            })?,
+                    );
                 }
                 FieldAttribute::Deprecation(attr) => {
                     output.deprecation = Some(attr);
@@ -752,7 +777,7 @@ impl GraphQLTypeDefiniton {
             let resolver_code = &variant.resolver_code;
 
             quote!(
-                Some(#variant_name) => Some(#resolver_code),
+                Some(#variant_name) => Ok(#resolver_code),
             )
         });
 
@@ -860,13 +885,14 @@ impl GraphQLTypeDefiniton {
             impl#impl_generics ::juniper::FromInputValue<#scalar> for #ty
                 #where_clause
             {
-                fn from_input_value(v: &::juniper::InputValue<#scalar>) -> Option<#ty>
-                {
-                    match v.as_enum_value().or_else(|| {
-                        v.as_string_value()
-                    }) {
+                type Error = ::std::string::String;
+
+                fn from_input_value(
+                    v: &::juniper::InputValue<#scalar>
+                ) -> Result<#ty, Self::Error> {
+                    match v.as_enum_value().or_else(|| v.as_string_value()) {
                         #( #from_inputs )*
-                        _ => None,
+                        _ => Err(format!("Unknown enum value: {}", v)),
                     }
                 }
             }
@@ -980,8 +1006,14 @@ impl GraphQLTypeDefiniton {
                     #field_ident: {
                         match obj.get(#field_name) {
                             #from_input_default
-                            Some(ref v) => ::juniper::FromInputValue::from_input_value(v)?,
-                            None => ::juniper::FromInputValue::<#scalar>::from_implicit_null()?,
+                            Some(ref v) => {
+                                ::juniper::FromInputValue::<#scalar>::from_input_value(v)
+                                    .map_err(::juniper::IntoFieldError::into_field_error)?
+                            },
+                            None => {
+                                ::juniper::FromInputValue::<#scalar>::from_implicit_null()
+                                    .map_err(::juniper::IntoFieldError::into_field_error)?
+                            },
                         }
                     },
                 )
@@ -1096,10 +1128,17 @@ impl GraphQLTypeDefiniton {
             impl#impl_generics ::juniper::FromInputValue<#scalar> for #ty #type_generics_tokens
                 #where_clause
             {
-                fn from_input_value(value: &::juniper::InputValue<#scalar>) -> Option<Self>
-                {
-                    let obj = value.to_object_value()?;
-                    Some(#ty {
+                type Error = ::juniper::FieldError<#scalar>;
+
+                fn from_input_value(
+                    value: &::juniper::InputValue<#scalar>
+                ) -> Result<Self, Self::Error> {
+                    let obj = value
+                        .to_object_value()
+                        .ok_or_else(|| ::juniper::FieldError::<#scalar>::from(
+                            format!("Expected input object, found: {}", value))
+                        )?;
+                    Ok(#ty {
                         #( #from_inputs )*
                     })
                 }